requirements.txt
streamlit
ultralytics
opencv-python
numpy
Pillow
gtts
pygame
